requirements.txt
# Incluye los modulos que requiere la app:Flask>=3.1.1
Flask>=3.1.2, <4
gunicorn>=23.0.0, <24
gevent>=25.9.1, <26
PyMySQL>=1.1.2, <2
Flask-SQLAlchemy>=3.1.1, <4
flask-marshmallow>=1.3.0, <2
marshmallow-sqlalchemy>=1.4.2, <2
python-dotenv>=1.1.1, <2
Flask-Limiter>=4.0.0, <5
flask-cors>=6.0.1, <7